fix: correct paths due to move of academy folder
