+++ frontend/src/components/buildings/academy/CreateTroopButton.js
@@ -6,8 +6,8 @@ import PropTypes from 'prop-types';
 import './CreateTroopButton.css';
 import { addTroopAction } from '../../../actions/TroopsActions';
 import { setErrorAction } from '../../../actions/ErrorActions';
-import Troop from '../../assets/troops/addTroop-removebg-preview.png';
-import Gold from '../../assets/sources/GoldIcon.svg';
+import Troop from '../../../assets/troops/addTroop-removebg-preview.png';
+import Gold from '../../../assets/sources/GoldIcon.svg';
 
 function CreateTroopButton({
   kingdom,
